Revert "Implementing gulp as task runner"
This reverts commit 351cfd0c010f9b72ca3f514309c0c60ff5a81709.
